requirements.txt
Django==2.1.15
djangorestframework==3.11.0
pyspark==2.4.4
redis==3.3.11